requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
pandas>=2.1.0
numpy>=1.23.0
scikit-learn>=1.3.0
scipy>=1.10.0
pyyaml>=6.0
tqdm>=4.65.0
flask>=3.0.0
gunicorn>=21.0.0
gymnasium>=0.29.0
pyarrow>=12.0.0
matplotlib>=3.7.0
uvicorn>=0.23.0
